style: copy getTableSize
