+++ sqle/driver/mysql/context/context.go
@@ -7,6 +7,7 @@ import (
 	"github.com/actiontech/sqle/sqle/driver/mysql/executor"
 	"github.com/actiontech/sqle/sqle/driver/mysql/util"
 	"github.com/pingcap/parser/ast"
+	"github.com/pkg/errors"
 )
 
 type TableInfo struct {
@@ -571,4 +572,29 @@ func (c *Context) GetSchemaEngine(stmt *ast.TableName, schemaName string) (strin
 		schema.engineLoad = true
 	}
 	return engine, nil
+}
+
+// getTableSize get table size.
+func (i *Inspect) getTableSize(stmt *ast.TableName) (float64, error) {
+	exist, err := i.Ctx.IsTableExist(stmt)
+	if err != nil {
+		return 0, errors.Wrapf(err, "check table exist when get table size")
+	}
+	if !exist {
+		return 0, nil
+	}
+
+	info, _ := i.Ctx.GetTableInfo(stmt)
+	if !info.sizeLoad {
+		conn, err := i.getDbConn()
+		if err != nil {
+			return 0, err
+		}
+		size, err := conn.ShowTableSizeMB(i.Ctx.GetSchemaName(stmt), stmt.Name.String())
+		if err != nil {
+			return 0, err
+		}
+		info.Size = size
+	}
+	return info.Size, nil
 }
